Update __init__.py
Major rewrite to incorporate continuity between consecutive scenes and events - added edges between pertinent nodes. Set tags to "location" and "event" for episode outlines and "speaker" and "speech" for manual transcripts. Add nodes at start and end of episodes. The start node of the first location is connected to the episode start node and the stop node of the last location is connected to the episode stop node.